# docs/swagger.yaml
...
info:
  contact:
    name: AAStar Support
    url: https://aastar.xyz
paths:
...
  /api/account/v1/sync:
    get:
      consumes:
      - application/json
      description: download accounts
      parameters:
      - description: how many accounts to download
        in: query
        name: count
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
      security:
      - JWT: []
      tags:
      - Account
...
securityDefinitions:
  JWT:
    description: Type 'Bearer \<TOKEN\>' to correctly set the AccessToken
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"